+++ lib/entities/inventory_item.dart
